pyproject.toml
[tool.poetry]
name = "postm"
version = "0.1.0"
description = ""
authors = ["pab-h <dev.pab.2020@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
pymongo = "^4.7.2"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
start = 'scripts:start'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
